+++ CHANGELOG.md
@@ -16,6 +16,9 @@
   * Initial release.
 * AADSocialIdentityProvider
   * Fixed missing permissions in settings.json
+* IntuneEndpointDetectionAndResponsePolicyWindows10
+  * Fixes an issue with `AutoFromConnector` as the Configuration package type.
+    FIXES [#5246](https://github.com/microsoft/Microsoft365DSC/issues/5246)
 * Intune workload
   * Fixed missing permissions in settings.json
 * SentinelAlertRule